Moved the generated bindings into a "sys" package
This makes it a lot easier to import the generated bindings into the make package since cargo does not have to consider the build script when building the package.
Also, this makes it possible to recompile the main package on save in Visual Studio Code (and surely other IDEs) without overloading the computer's resources.
Also, this makes it possible to easily publish the crate without disabling checks!