removed all the get_desc functions
in my opinion, they are a bit useless since all the objects are created with known values to the user.
The only one that wasn't removed is the swap chain. For now it's used in the tutorials and may be removed in a near future if a better solution is implemented.