SwapChainDesc is now transparent
instead of copying every feild of SwapChainDesc into its ported version, SwapChainDesc is now repr(transparent).
It has the same ABI as the a pointer to the desc, and values are accesses through dereferencing the pointer.
A phantom data was added to the transparent struct to prevent the desc from outliving the swap chain itself.